backend/requirements.txt
fastapi
uvicorn
pydantic
openai
sympy
numpy
python-dotenv
httpx
